Add Loading page
This is necessary because Auth.currentUser() returns a Promise. When a
user is logged in and the app is reloaded, there is a brief flicker
during which the SignIn screen is shown before the authenticated app
appears. This Loading screen replaces that flicker and will toggle to
either signIn or signedIn, as appropriate.