Adding ability to specify OAuth/Federation clients the React app supports
The `withFederated()` HoC now supports an optional second argument that is
an opts object that allows us to specify exactly which SSO/OAuth clients
we want to support within an application. By default, all three (Amazon, Google and Facebook)
are mixed in. We can opt-in specifically by passing in an object with the
services we want to support. For example, using both Amazon and Facebook (excluding Google):

```javascript
const Comp = withFederated(MockComp, { amazon: true, facebook: true });
```

This commit also upgraded Jest to the current 22.x version, Enzyme 3.1.x and added
in jest-enzyme to the dev dependencies. I have found the jest-enzyme matchers
(https://github.com/FormidableLabs/enzyme-matchers) to be very convenient when
testing React applications and components. If you want me to rollback this
addition I can, and will update the config/tests.

Fixes #554