feat(aws-amplify-angular): provide singletone amplify service
* Provide the amplify service through a
factory provider in the angular module.
Currently, consumers of the library are obligated
to provide the service at root module level.
With this change, this is no longer required.

Furthermore, its ensured that the service is singletone
even when the aws module is imported by lazy modules
or imported/exported by shared ones.

Sidenote: the docs currently state that registering
a provider for the service is an optional step
isnt accurate, as angular will throw an InjectionError
when there is no provider for the service registered.