remove import of CSS files in JSX
The css can be prepackaged and explicitly imported via Webpack or
another way. The explicit inclusion of css styles requires a CSS
processor in dependant aplications that use the amplify-react library.
This PR removes that hard dependency for people that do not wish to use
the CSS processor (e.g. CSS in JS)