In app messaging/remove react native UI (#10425)
<!--
Please make sure to read the Pull Request Guidelines:

https://github.com/aws-amplify/amplify-js/blob/main/CONTRIBUTING.md#pull-requests
-->

#### Description of changes
<!--
Thank you for your Pull Request! Please provide a description above and
review
the requirements below.
-->

This PR restores the `aws-amplify-react-native` package to its state on
the `main` branch. All relevant changes have been migrated to Amplify UI
and this package should not be used with in-app messaging.

Also pinned the version of the pinpoint client used in `notifications`
to a specific targeted version. Latest version of AWS SDK is currently
resulting in build failures due to webpack incompatibility. This can
addressed at some later time across the JS monorepo.

#### Description of how you validated changes
`yarn test`

#### Checklist
<!-- Remove items that do not apply. For completed items, change [ ] to
[x]. -->

- [x] PR description included
- [x] `yarn test` passes

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Co-authored-by: Chris Fang <chrfang@amazon.com>