fix(auth): session debouncer loads empty signinUserSession
This change fixes a race condition where inflightSessionPromise is set
to null before returning.

This change also fix the issue where concurrent call to currentAuthenticatedUser
returns users with empty signinUserSession. This is because in the userSession
call debouncer skip the consequent call to user.getSession, so the user instance
contains empty session property.