Adjust event message size
The substring was previously running to the eventSize so not cutting down the message. Changed to set a hard limit of max event size with its buffer. Restated eventsize for downstream logic.

Ideally it should be changed to split a large `message` into chunks of max size and submit multiple events.