fix(core): add @types/node-fetch to runtime dependency
The isomorphic-unfetch dependency re-exports artifacts from node-fetch
package in its .d.ts file. However, the node-fetch package does not
contain any .d.ts file with itself. So the exported artifacts can not be
resolved by tsc. It's not an issue before possibly because the client-s3
contains transitive runtime dependency of @types/node-fetch. Since
it's removed now, we are encountered with this error.

Adding the dependency to runtime instead of dev or peer deps to prevent
brokage for customers